requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
httpx==0.28.1
mercadopago==2.2.3
Pillow==11.0.0
python-slugify==8.0.4
emails==0.6
jinja2==3.1.4
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
aiofiles==24.1.0
redis==5.2.1
celery==5.4.0
flower==2.0.1
